refactor : 불필요 class 제거
