+++ src/main/java/core/jdbc/JdbcTemplate.java
@@ -5,45 +5,74 @@ import core.exception.JdbcException;
 import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
 
 public class JdbcTemplate {
-    private final DataSource dataSource;
+	private final DataSource dataSource;
 
-    public JdbcTemplate(DataSource dataSource) {
-        this.dataSource = dataSource;
-    }
+	public JdbcTemplate(DataSource dataSource) {
+		this.dataSource = dataSource;
+	}
 
-    public <T> List<T> selectList(String sql, Object[] args, int[] argTypes, RowMapper<T> rowMapper) {
-        return execute(sql, new SelectQueryAction(), new ParameterSetter(args, argTypes), new ListRowMapper<T>(rowMapper));
-    }
+	public <T> List<T> selectList(String sql, Object[] args, int[] argTypes, RowMapper<T> rowMapper) {
+		return execute(sql, new SelectQueryAction(), new ParameterSetter(args, argTypes), (rs) -> {
+				List<T> tempList = new ArrayList<>();
+				while(rs.next()) {
+					tempList.add(rowMapper.resultMapping(rs));
+				}
+				return tempList;
+			}
+		);
+	}
 
-    public <T> T select(String sql, Object[] args, int[] argTypes, RowMapper<T> rowMapper) {
-        return execute(sql, new SelectQueryAction(), new ParameterSetter(args, argTypes), new SingleRowMapper<T>(rowMapper));
-    }
+	public <T> T select(String sql, Object[] args, int[] argTypes, RowMapper<T> rowMapper) {
+		return execute(sql, new SelectQueryAction(), new ParameterSetter(args, argTypes), (rs) -> {
+				T value = null;
 
-    public int update(String sql, Object[] args, int[] argTypes) {
-        return execute(sql, new UpdateQueryAction(), new ParameterSetter(args, argTypes), new RowCountMapper());
-    }
+		        while(rs.next()) {
 
-    private<T, R> R execute(String sql, QueryAction<T> queryAction, ParameterSetter parameterSetter, ResultMapper<T, R> resultMapper) {
-        try(Connection connection = this.dataSource.getConnection();
-            PreparedStatement preparedStatement = connection.prepareStatement(sql)
-            ) {
+		            if(value != null) {
+		                throw new SQLException("has multipleRows!!!");
+		            }
 
-            setArgsIfSupport(preparedStatement, parameterSetter);
-            T result = queryAction.action(preparedStatement);
+		            value = rowMapper.resultMapping(rs);
+		        }
 
-            return resultMapper.resultMapping(result);
-        } catch (SQLException e) {
-            throw new JdbcException(e);
-        }
-    }
+		        return value;
+			}
+		);
+	}
 
-    private void setArgsIfSupport(PreparedStatement preparedStatement, ParameterSetter parameterSetter) throws SQLException {
-        if(parameterSetter != null) {
-            parameterSetter.setArgs(preparedStatement);
-        }
-    }
+	public int update(String sql, Object[] args, int[] argTypes) {
+		return execute(sql, new UpdateQueryAction(), new ParameterSetter(args, argTypes), new ResultMapper<Integer, Integer>() {
+
+			@Override
+			public Integer resultMapping(Integer result) throws SQLException {
+				return result;
+			}
+		});
+	}
+
+	private<T, R> R execute(String sql, QueryAction<T> queryAction, ParameterSetter parameterSetter, ResultMapper<T, R> resultMapper) {
+		try(Connection connection = this.dataSource.getConnection();
+				PreparedStatement preparedStatement = connection.prepareStatement(sql)
+				) {
+
+			setArgsIfSupport(preparedStatement, parameterSetter);
+			T result = queryAction.action(preparedStatement);
+
+			return resultMapper.resultMapping(result);
+		} catch (SQLException e) {
+			throw new JdbcException(e);
+		}
+	}
+
+	private void setArgsIfSupport(PreparedStatement preparedStatement, ParameterSetter parameterSetter) throws SQLException {
+		if(parameterSetter != null) {
+			parameterSetter.setArgs(preparedStatement);
+		}
+	}
 }
